modify Introduction.py

+++ Introduction.py
@@ -13,14 +13,14 @@ pre_path = "./data/pre_run/"
 
 def load_raw_data(s, path=pre_path):
 
-    pai = np.load(path + f"pai_reduced_{s}.npy")
+    pai = np.load(path + f"pai_reduced_{s}.npy").astype(np.float16)
     us = np.load(path + f"us_{s}.npy")
 
     return pai, us
 
 def pre_processing(path=pre_path):
 
-    pai_ready = np.load(path + f"pai_ready_reduced_{s}.npy")
+    pai_ready = np.load(path + f"pai_ready_reduced_{s}.npy").astype(np.float16)
     us_ready = np.load(path + f"us_ready_{s}.npy")
     st.success('Data has been processed.')
